refactor(reports): extract summary report to reports_summary.go
